vrp/app/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
SQLAlchemy==1.2.12
Werkzeug==0.14.1
